+++ NEWS.md
@@ -5,6 +5,10 @@
 * `across()` now errors if the argument `.cols` is not provided (either named or
   unnamed). This behavior was deprecated in `dplyr` 1.1.0.
   
+**New features**
+
+* `summarize()` now works on ungrouped data and returns a 1-row output.
+  
 **Bug fixes**
 
 * `summarize()` no longer errors when `across(everything(), ...)` is used with